Track Cargo.lock
There is a weird issue happened occasionally. That in src/bin/main.rs,
`target` in line 144 cannot be identified as `String`
and `chunk_by` in line 122 is not recognized.

The issue seemed to be fixed after `cargo update` but happened later. Hence tracking the Cargo.lock
and see if there was a case to cause cargo resolving to a wrong version of deps.